merged

+++ include/if3_util.hpp
@@ -55,7 +55,7 @@ namespace if3
       if (a.length() == strlen(b))
       {
         r = true;
-        for (int i = 0; i < a.length(); ++i)
+        for (size_t i = 0; i < a.length(); ++i)
         {
 
           if (toupper(a[i]) != toupper(b[i]))
